requirements.txt
ultralytics
clearml
pyyaml
opencv-python
torch
torchvision
onnx
onnxruntime
tabulate
termcolor
ruamel.yaml
terminaltables
blessed
rich
super-gradients